cmake: Allow specifying several libs in sysrepo.pc
As a side effect of pull request sysrepo/sysrepo#389, sysrepo.pc
pkg-config file is changed so that it now specifies both libsysrepo.so
and libsysrepo-engine.so for linking.  This patch adapts to that change
(if it gets merged) so that netopeer2 now supports linking with sysrepo
again.

I am not completely happy about this. I asked on #cmake IRC channel and
apparently The Modern Way™ would be for sysrepo to install a special
fancy cmake file which one could use to bypass pkg-config altogether. I
asked because I got puzzled by the suggested dance to take pkg-config's
results and call cmake's find_library on top of them.

Another alternative would be to add link_directories(), but that would
require calling find_module(SYSREPO) prior to defining targets -- please
indicate whether you prefer this LIST() dance, or whether you're OK with
moving this order to detect libs at first, and only then define targets.