cmake: Use GNUInstallDirs for sysconf (aka /etc)
It makes sense to install everything, including the configuration, to
one directory during development or when using prefixed installation for
some reason. However, when the installation is being performed "for
real" for production use, then the configuration files do not belong to
/usr/etc, but straight to /etc.

It turns out that the GNUInstallDirs CMake module already contains
logic for this. By default, a CMAKE_INSTALL_PREFIX=/usr/local is still
used which means that the etc prefix is /usr/local/etc/. When installing
with CMAKE_INSTALL_PREFIX=/usr, though, this gets translated to just
/etc, which is exactly what I wanted to achieve in my original version
of this patch.