server BUGFIX thread starving when processing sysrepo's changes in context
When changing context due to the changes in sysrepo's context (removing/adding
schemas), the thread performing this change (need write lock on context) starved
by other threads processing the client requests (need read lock on context).

Already the previous commit should help to change timeouts when waiting for
client requests. However, there can still be cases (mainly when the number of
worker threads is increased) that it actually does not help. Therefore now
we change the attribute of the rwlock to prefer write locks instead of (default)
read locks causing starving. This should be complete solution, unfortunately it
is not portable (using pthread_rwlockattr_setkind_np()) - the issue is described
in the added KNOWNISSUES file.

Fixes #139