Fix enabling feat depending on later schema feat
`np2srv_init_schemas()` walks through schemas in the order they are returned
from sysrepo via `np2srv_sr_list_schemas()` and attemps to enable features
that sysrepo indicates are supposed to be enabled.

However, if a feature in schema `n` depends on a feature that is found in
schema `n + 1` (or any subsequent schema, for that matter), enabling the
feature in schema `n` will fail because the feature in schema `n + 1`
has not yet been enabled.

The additional unit test included in this commit demonstrates the problem.

This commit implements a solution as follows:
1. When walking through the schemas returned from sysrepo, keep a list
   containing all the features that will need to be enabled.
2. Now walk through all the schemas attempting to enable the recorded features.
   Some features may fail to enable due to depending on other features that
   have not yet been enabled. So continue walking through the schemas
   attempting to enable features as long as (1) we are able to make progress
   and enable at least one new feature, or (2) there are no more features to
   enable.

This approach should allow us to enable all of the required features
regardless of how far apart they are on a dependency graph.