Fix longer feature dep chains
This is a follow-up to pull request #248, which fixed a problem enabling a
feature that depended on a feature in a subsequent schema that had not yet
been enabled.

The follow-up optimizations in commit 6cb1612 inadvertently changed the
desired behavior, in that a feature dependency chain more than two features
deep will no longer be enabled properly. The problem is that the
`making_progress` loop in main.c line 876 is never entered.

I've included a change to the unit test from #248 that demonstrates a
feature dependency chain three features deep, instead of the original test
which was only two features deep.