Add an option for overriding module permissions
The current build hardcodes specific options to be passed to
`sysrepoctl` at installation time which in turn requires running as
root. This is a problem when cross-building under tools such as
Buildroot which handle root access in another way.

This patch adds an option to cmake to modify these security settings. A
build recipe can enforce the file permission later. I am doing just that
in the Buildroot.