keystored: enable SSH key provisioning by an external script
Our build scripts (Buildroot) generate a fileysstem image to be used by
many devices, which is why they do not contain any key material. This
means that by default, netopper2-keystored won't install any keys, and
that the netopeer2-server's installation fails because its default
configuration refers to a non-existing SSH key:

  -- Importing module ietf-system into sysrepo...
  CMake Error at cmake_install.cmake:119 (message):
      Command sysrepocfg import failed:
      [ERR] (rp_dt_commit:864) Validation after merging failed
      [ERR] (sr_commit:2089) Commit operation failed with 1 error(s).
      [ERR] (srcfg_import_datastore:911) Error returned from sr_commit: Validation of the changes failed.
      [ERR] (srcfg_import_datastore:914) /ietf-netconf-server:netconf-server/listen/endpoint[name='test_ssh_listen_endpt']/ssh/host-keys/host-key[name='test_ssh_listen_key']/public-key : Leafref "/ietf-keystore:keystore/ietf-keystore:private-keys/ietf-keystore:private-key/ietf-keystore:name" of value "ssh_host_rsa_key" points to a non-existing leaf.

Without this patch, I would have to provide full configuration for the
NETCONF server and for the keystored daemon. My preference is to avoid
duplicating the (ricky) logic which is already available in these cmake
installation scripts.

Therefore, this patch simply adds yet another flag which means that an
external deploy script will provide a PEM-encoded SSH host file at a
well-known location. That is pretty easy to provide via a systemd unit
file in a similar manner to how the OpenSSH server already pre-generates
its keys if they are not available.