cmake: split installation tasks into scripts
This commit takes all the installation tasks (yang model installation, ssh key
import) out of cmake, in separate scripts, and calls those scripts during the
install target. It makes the code more readable.

MODEL_INSTALL and SSH_KEY_INSTALL options can be set to off to avoid the
scripts to be called in keystored makefile.

The yang model installation script is only called when ENABLE_CONFIGURATION is
set in the server makefile.

KEYSTORED_DEFER_SSH_KEY is removed, as it is replaced by SSH_KEY_INSTALL.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>