Unit test for selection node filter on key
This commit adds a unit test that demonstrates a selection node subtree
filter, where the selection node is a key.

In netopeer2@9e793f1, `opget_build_subtree_from_sysrepo()` was modified to
call `opget_build_subtree_from_sysrepo_is_key()`. The side-effect is that
nodes that are keys are no longer added to the subtree being built up as a
response to the get RPC. An empty subtree will be erroneously returned,
instead of a subtree with the matching key node(s).