Add flag to treat SIGHUP/SIGUSR1 as termination
`netopeer2-server`'s signal handler currently attempts to restart
netopeer2-server if it receives `SIGHUP` or `SIGUSR1`. Unfortunately, we've
noticed several problems with the server restart mechanism. For example,
libssh's thread.c does not properly clean up its state, so netopeer2-server
crashes with a segmentation fault when restarting.

This pull request does not attempt to address any of the problems with the
server restart mechanism. Instead, it adds an optional configuration flag
`DISABLE_RESTART_SIGNALS` that allows us to treat `SIGHUP` and `SIGUSR1` as
signals that will _terminate_ the server instead of restarting it. The flag
is disabled by default, so the existing behavior remains unchanged.