Fix issue with actions from augments
Before this change an action added by an augment would always be
treated as "unimplemented", even if there is a subscriber for it
downstream of sysrepo.

This patch causes the failing test cases (see [1] and [2]) to start
passing. This was not added to a unit test as main.c isn't being
tested in that context right now.

[1]: https://travis-ci.org/ADTRAN/netopeer2-integration-tests/builds/413288159#L7278
[2]: https://github.com/ADTRAN/netopeer2-integration-tests/blob/master/tests/test_action.py#L58