Cache ly_ctx_info to improve get performance
Fixes #304

This change results in a 3x to 4x speedup for unfiltered get requests (or
requests for the ietf-yang-library contents).

We now cache the results of the `ly_ctx_info(np2srv.ly_ctx)` call and save it
in the `np2srv` global, to avoid repeating the call when it would just return
the same data.

We update the cached value when necessary, namely, when a module or feature
is changed.

Because the contents of the cache should change relatively infrequently, it
seems worth it to pay the cost early instead of for every request.