Test to demonstrate container with default leaf
This commit changes one of the test models such that it has a container with
a leaf with a default. Prior to the cache optimizations in 6ab0617, an
empty container whose leafs were all default values was suppressed and
not generated as part of the running config. After 6ab0617, an empty
container is now presented with the running-config.

Thus `test_edit_delete1()` in `test_edit_get_config` will fail because the
following empty container will be presented with the configuration:

```
<test-container xmlns=urn:ietf:params:xml:ns:yang:test-feature-c>
</test-container>
```

The tests pass in commits prior to 6ab0617.