Require sequential consistency when synchronizing via atomics
When used with the relaxed memory order, atomic variables are not a
synchronization primitive:

> Atomic operations tagged `memory_order_relaxed` are not
> synchronization operations; they do not impose an order among
> concurrent memory accesses. They only guarantee atomicity and
> modification order consistency.

-- https://en.cppreference.com/w/c/atomic/memory_order#Relaxed_ordering

We actually want to use these variables as synchronization points,
guaranteeing that everything which "has happened" (in the happens-before
terminology) within thread A is also visible in thread B. That's
synchronization of *other* memory, and just an atomic inc/dec of the
flag variable on its own is not enough.

This reverts commit 579ad388b67979f2557d799893afcabdd44493c0 (with a
manual fix-up to fix build).

Cc: #405