server: disable cwd yang search dir
When creating the libyang context, disable searching for yang modules in
the current working directory. Depending on where netopeer2-server is
started, we can get surprising warnings/errors:

For example, when starting the server from "/":

  Reading module "ietf-yang-library".
  Searching for "ietf-yang-types" in /.
  Unable to get information about "initrd.img.old" file in "/" when
  searching for (sub)modules (No such file or directory)
  Unable to get information about "vmlinuz.old" file in "/" when
  searching for (sub)modules (No such file or directory)
  ...
  Module "ietf-yang-library@2019-01-04" successfully parsed as implemented.

In that case, the "non-exsistant" items are symlinks:

  ~# ls -lhF /initrd* /vmlinuz*
  lrwxrwxrwx 1 root root 32 Aug 14 16:19 /initrd.img -> boot/initrd.img-5.0.0-20-generic
  lrwxrwxrwx 1 root root 32 Aug 14 16:35 /initrd.img.old -> boot/initrd.img-5.0.0-20-generic
  lrwxrwxrwx 1 root root 29 Aug 14 16:19 /vmlinuz -> boot/vmlinuz-5.0.0-20-generic
  lrwxrwxrwx 1 root root 29 Aug 14 16:35 /vmlinuz.old -> boot/vmlinuz-5.0.0-20-generic

I am not sure why we get this "No such file or directory" error but
I cannot find a good reason to search for yang modules in the current
directory anyway.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>