server BUGFIX do not abort endpoint changes processing on error
When changing the configuration to add new listen endpoints, if one of
these endpoints specifies an invalid parameter (for example,
a "local-address" which is not present on any interface on the system),
the nc_server_endpt_set_address() function will return an error (since
the socket cannot be created).

If such errors occur, do not abort the processing of changes since other
endpoint address/port may be valid and should be configured. Otherwise,
these changes are lost and will never be configured.

If one error occurred, return an error to notify the sysrepo client
which applied the configuration that something went wrong.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>