Allow moving list/leaf-list items to the last position
The code wanted to deliberately ignore any SR_MOVE_LAST values. I think
that this happened because of that C convention for adding a FOO_LAST
items to enums to signify their largest allowed value.

I did not want to modify sysrepo for this as well, so I just added a
random garbage sentinel for this.

Bug: https://github.com/CESNET/netopeer2/issues/636