BUGFIX last position move (#639)
* Allow moving list/leaf-list items to the last position

The code wanted to deliberately ignore any SR_MOVE_LAST values. I think
that this happened because of that C convention for adding a FOO_LAST
items to enums to signify their largest allowed value.

I did not want to modify sysrepo for this as well, so I just added a
random garbage sentinel for this.

* Do not reset relative list XPath too early

When processing YANG list moves, the code would correctly parse the
relative XPath in the edit_get_move(), but before it could be consumed
via `np2srv_sr_move_item`, it got reset when the main iteration loop
processes the next change item. As a result, a relative move operation
was always attempted with a nullptr for the relative item XPath.

Fix this by only resetting the `rel` when encountering a new leaf-list
or a new list. I have no clue if this is actually correct (what about
nested lists, for example?). However, my C skills are too limited to
correctly follow constructs with gotos into the middle of if statements,
so this is my best attempt -- sorry.

It seems that there are no tests for list moves, so I'm not modifying
them.

Fixes #636