BUGFIX relative move operation
Relative move operation are made with the use of the `yang:key`
attribute. This attribute specifies the relative instance for the
operation. The instance keys in the attribute are prefixed with the YANG
module prefix and that is what new libyang returns (old libyang returned
module names). Sysrepo can't handle prefixes. However, it can handle
non-prefixed key names, so this patch parses the `yang:key` string to
remove the prefix.

Fixes #641