tests: use relative paths for Unix sockets
Linux limits the socket paths to about 108 characters, and my setup uses
rather long prefixes for the build dir. Since a relative path works here
just fine, use that.

It is possible to set the CWD of a test run via CMake's
WORKING_DIRECTORY property, but the default works just fine here, so
there's no need to tweak that. There's a potential regression if a
developer runs the test executable directly, but doing that bypasses a
lot of CMake/CTest features already, such as fixtures or environment
variables which are already used by this project, so I think this is a
safe fix.

Fixes: https://github.com/CESNET/netopeer2/issues/1020