np_test UPDATE set ln2 mod dir env &  default cfg
Sets the env variable from which the modules will be loaded in a test.
No longer wait for a socket to know that server is ready, instead
wait for its pidfile.
Create default UNIX socket manually, since the -U opt was removed from
np2-srv as it's no longer needed.