main BUGFIX fix empty rpc-reply
When an RPC output has a container and the RPC callback returns nothing,
the np2srv_rpc_cb does not add the 'ok' empty leaf. The rpc-reply is
empty.
The client receives an error when it reads the RPC response with
nc_recv_reply, because it expects either 'ok' or 'data'.
Look at the content of the reply and add the 'ok' leaf if no child with
the default flag is found.

Fixes: 3936132d91d6 ("server UPDATE call NC RPC directly, not using sysrepo")
Link: https://datatracker.ietf.org/doc/html/rfc7950#section-7.14.4
Link: https://datatracker.ietf.org/doc/html/rfc6241#section-4.4